// packages/test/generatedCode/policies/permissionGrantPolicies/item/permissionGrantPolicyItemRequestBuilderGetRequestConfiguration.ts
import {PermissionGrantPolicyItemRequestBuilderGetQueryParameters} from './permissionGrantPolicyItemRequestBuilderGetQueryParameters';
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface PermissionGrantPolicyItemRequestBuilderGetRequestConfiguration {
    /** Request headers */
    headers?: Record<string, string>;
    /** Request options */
    options?: RequestOption[];
    /** Request query parameters */
    queryParameters?: PermissionGrantPolicyItemRequestBuilderGetQueryParameters;
}